[AutoDeploy] Phi3/4, position_id, HF loading, parameter aliasing   (#25)
Includes #17 , #18, #21 , #23

* move Phi PR for sprint


* remove Phi3 patch and unfuse_weight transformation


* update Phi3/4 patches


* minor


* minor


* Update the position id to int64


* Add support for handling aliased parameters


* update tests


* lint


* lint


* removed load state dict with assign context manager


* formatting fix


* Removed related unit tests


* [AutoDeploy] correctly identifying how to load HF checkpoint weights


* [AutoDeploy] correctly load sharded weights


* skip sharding if num_users>1


---------

+++ tensorrt_llm/_torch/auto_deploy/custom_ops/attention_interface.py
@@ -78,7 +78,7 @@ class SequenceInfo:
     ## [UPDATE WITH CARE] TENSOR FIELDS THAT WILL BE PASSED TO PREPARE_METADATA OP #################
     # input_ids MUST ALWAYS BE THE FIRST FIELD
     input_ids: torch.Tensor = field(default_factory=lambda: torch.zeros(1, 1, dtype=torch.int))
-    position_ids: torch.Tensor = field(default_factory=lambda: torch.zeros(1, 1, dtype=torch.int))
+    position_ids: torch.Tensor = field(default_factory=lambda: torch.zeros(1, 1, dtype=torch.long))
 
     seq_len: torch.Tensor = field(default_factory=lambda: torch.ones(1, dtype=torch.int))
     input_pos: torch.Tensor = field(default_factory=lambda: torch.zeros(1, dtype=torch.int))
@@ -106,7 +106,7 @@ class SequenceInfo:
         total_tokens = min(self.max_num_tokens, self.max_batch_size * max_seq_len_adjusted)
         self._num_pages = (total_tokens) // self.page_size + (total_tokens % self.page_size > 0)
         self.input_ids = torch.ones(self.max_batch_size, 1, dtype=torch.int)
-        self.position_ids = torch.zeros(self.max_batch_size, 1, dtype=torch.int)
+        self.position_ids = torch.zeros(self.max_batch_size, 1, dtype=torch.long)
         self.seq_len = torch.empty(self.max_batch_size, dtype=torch.int)
         self.input_pos = torch.empty_like(self.seq_len)
         self.cache_loc = torch.empty(self.num_pages, dtype=torch.int)
@@ -341,7 +341,7 @@ class SequenceInfo:
     def _update_position_ids(self) -> None:
         # set new position_ids as new tensor from input_pos and seq_len via torch.arange
         position_ids_list = [
-            torch.arange(in_pos, in_pos + seq_len, dtype=torch.int)
+            torch.arange(in_pos, in_pos + seq_len, dtype=torch.long)
             for in_pos, seq_len in zip(self.input_positions, self.sequence_lengths)
         ]
         self.position_ids = torch.cat(position_ids_list, dim=0).to(self.device)
